Fix for Issue #10
Bug description: After starting the code with Step, the CPU goes to
100%, until Stop is pressed
Reason: In step mode the execution stucks inside a loop in the
execute_command function and continuosly updates canvas until the stop
or run buttons get pressed.
Solution: Call execute_command every time when the user presses the step
button and the loop works only in run mode.